# api/openapi/platform-v1.yaml
components:
  schemas:
...
    ErrorEnvelope:
      additionalProperties: false
      properties:
        code:
          enum:
          - INVALID_ARGUMENT
          - UNAUTHENTICATED
          - FORBIDDEN
          - NOT_FOUND
          - CONFLICT
          - PRECONDITION_FAILED
          - RATE_LIMITED
          - SOURCE_DEGRADED
          - GRAPH_NOT_READY
          - GRAPH_PARTIAL
          - TIMEOUT
          - BUDGET_EXHAUSTED
          - POLICY_DENIED
          - RISK_ACK_REQUIRED
          - RISK_ACK_COMMAND_MISMATCH
          - STEP_UP_REQUIRED
          - STEP_UP_EXPIRED
          - IDEMPOTENCY_CONFLICT
          - INTERNAL
          - CAPABILITY_DISABLED
          - SOURCE_NOT_CONFIGURED
          - IDEMPOTENCY_IN_PROGRESS
          - IDEMPOTENCY_KEY_REUSED
          - EXECUTION_PROFILE_DENIED
          - COMMAND_POLICY_DENIED
          - EXECUTION_UNKNOWN
          - OUTPUT_TRUNCATED
          - BUNDLE_SIGNATURE_INVALID
          - BUNDLE_DIGEST_MISMATCH
          - BUNDLE_EXTRACTION_REJECTED
          - PROFILE_COMPONENT_CONFLICT
          - UPSTREAM_COMPATIBILITY_UNVERIFIED
          - OPENBAO_SEALED
          - HOST_NOT_ONBOARDED
          type: string
        details:
          additionalProperties: true
          type: object
        message:
          type: string
        requestId:
          $ref: '#/components/schemas/Identifier'
        retryable:
          type: boolean
      required:
      - code
      - message
      - requestId
      - retryable
      type: object
...
    Identifier:
      format: identifier
      type: string
...
    SP04ErrorEnvelopeV2:
      additionalProperties: false
      properties:
        code:
          enum:
          - INVALID_ARGUMENT
          - UNAUTHENTICATED
          - FORBIDDEN
          - NOT_FOUND
          - CONFLICT
          - PRECONDITION_FAILED
          - RATE_LIMITED
          - SOURCE_DEGRADED
          - GRAPH_NOT_READY
          - GRAPH_PARTIAL
          - TIMEOUT
          - BUDGET_EXHAUSTED
          - POLICY_DENIED
          - RISK_ACK_REQUIRED
          - RISK_ACK_COMMAND_MISMATCH
          - STEP_UP_REQUIRED
          - STEP_UP_EXPIRED
          - IDEMPOTENCY_CONFLICT
          - INTERNAL
          - CAPABILITY_DISABLED
          - SOURCE_NOT_CONFIGURED
          - IDEMPOTENCY_IN_PROGRESS
          - IDEMPOTENCY_KEY_REUSED
          - EXECUTION_PROFILE_DENIED
          - COMMAND_POLICY_DENIED
          - EXECUTION_UNKNOWN
          - OUTPUT_TRUNCATED
          - BUNDLE_SIGNATURE_INVALID
          - BUNDLE_DIGEST_MISMATCH
          - BUNDLE_EXTRACTION_REJECTED
          - PROFILE_COMPONENT_CONFLICT
          - UPSTREAM_COMPATIBILITY_UNVERIFIED
          - OPENBAO_SEALED
          - HOST_NOT_ONBOARDED
          - SOURCE_SCOPE_UNVERIFIED
          - STALE_CONTEXT
          - CURSOR_EXPIRED
          type: string
        details:
          additionalProperties: true
          type: object
        message:
          type: string
        requestId:
          $ref: '#/components/schemas/Identifier'
        retryable:
          type: boolean
        schemaVersion:
          type: string
          enum:
          - error-envelope/v2
      required:
      - code
      - message
      - requestId
      - retryable
      - schemaVersion
      type: object
...
    InvestigationUsageV1:
      type: object
      properties:
        toolCalls:
          type: integer
          minimum: 0
        rawQueries:
          type: integer
          minimum: 0
        resultBytes:
          type: integer
          minimum: 0
        graphNodes:
          type: integer
          minimum: 0
        evidenceItems:
          type: integer
          minimum: 0
        modelRequests:
          type: integer
          minimum: 0
        inputTokens:
          type: integer
          minimum: 0
        outputTokens:
          type: integer
          minimum: 0
        modelCostMicros:
          type: integer
          minimum: 0
      required:
      - toolCalls
      - rawQueries
      - resultBytes
      - graphNodes
      - evidenceItems
      - modelRequests
      - inputTokens
      - outputTokens
      - modelCostMicros
      additionalProperties: false
    InvestigationBudgetV1:
      type: object
      properties:
        toolCalls:
          type: integer
          minimum: 0
        rawQueries:
          type: integer
          minimum: 0
        resultBytes:
          type: integer
          minimum: 0
        graphNodes:
          type: integer
          minimum: 0
        evidenceItems:
          type: integer
          minimum: 0
        modelRequests:
          type: integer
          minimum: 0
        inputTokens:
          type: integer
          minimum: 0
        outputTokens:
          type: integer
          minimum: 0
        modelCostMicros:
          type: integer
          minimum: 0
        durationSeconds:
          type: integer
          minimum: 1
          maximum: 600
        allowedDataClasses:
          type: array
          items:
            enum:
            - D0
            - D1
          minItems: 1
          maxItems: 2
          uniqueItems: true
      required:
      - toolCalls
      - rawQueries
      - resultBytes
      - graphNodes
      - evidenceItems
      - modelRequests
      - inputTokens
      - outputTokens
      - modelCostMicros
      - durationSeconds
      - allowedDataClasses
      additionalProperties: false
    InvestigationScopeV1:
      type: object
      properties:
        tenantId:
          type: string
          format: uuid
        clusterUid:
          type: string
          minLength: 1
          maxLength: 128
        namespaces:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 128
          maxItems: 1000
          uniqueItems: true
        resources:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 128
          maxItems: 1000
          uniqueItems: true
        clusterScoped:
          type: boolean
        authorizationRevision:
          type: string
          minLength: 1
          maxLength: 256
      required:
      - tenantId
      - clusterUid
      - namespaces
      - resources
      - clusterScoped
      - authorizationRevision
      additionalProperties: false
    InvestigationJobV2:
      type: object
      properties:
        schemaVersion:
          enum:
          - investigation-job/v2
        tenantId:
          type: string
          format: uuid
        jobId:
          type: string
          format: uuid
        incidentId:
          type: string
          format: uuid
        subject:
          type: string
          minLength: 1
          maxLength: 256
        triggerRevision:
          type: integer
          minimum: 1
        triggerKind:
          enum:
          - manual
          - policy
          - new_evidence
        policyVersion:
          type: string
          minLength: 1
          maxLength: 128
        scope:
          $ref: '#/components/schemas/InvestigationScopeV1'
        effectiveScopeDigest:
          type: string
          pattern: ^sha256:[a-f0-9]{64}$
        toolCatalogDigest:
          type: string
          pattern: ^sha256:[a-f0-9]{64}$
        state:
          enum:
          - queued
          - running
          - succeeded
          - partial
          - failed
          - cancelled
          - expired
        budget:
          $ref: '#/components/schemas/InvestigationBudgetV1'
        budgetReserved:
          $ref: '#/components/schemas/InvestigationUsageV1'
        budgetConsumed:
          $ref: '#/components/schemas/InvestigationUsageV1'
        expiresAt:
          type: string
          format: date-time
        eventSeq:
          type: integer
          minimum: 0
        nextCallSeq:
          type: integer
          minimum: 0
        result:
          $ref: '#/components/schemas/InvestigationResultV1'
        errorCode:
          type: string
          minLength: 1
          maxLength: 128
      required:
      - schemaVersion
      - tenantId
      - jobId
      - incidentId
      - subject
      - triggerRevision
      - triggerKind
      - policyVersion
      - scope
      - effectiveScopeDigest
      - toolCatalogDigest
      - state
      - budget
      - budgetReserved
      - budgetConsumed
      - expiresAt
      - eventSeq
      - nextCallSeq
      additionalProperties: false
    InvestigationStepV2:
      type: object
      properties:
        stepId:
          type: string
          format: uuid
        toolName:
          type: string
          minLength: 1
          maxLength: 128
        argsDigest:
          type: string
          pattern: ^sha256:[a-f0-9]{64}$
        state:
          enum:
          - running
          - succeeded
          - failed
        result: {}
        errorCode:
          type: string
          maxLength: 128
      required:
      - stepId
      - toolName
      - argsDigest
      - state
      additionalProperties: false
    InvestigationResultV1:
      type: object
      additionalProperties: false
      properties:
        schemaVersion:
          type: string
          enum:
          - investigation-result/v1
        status:
          type: string
          enum:
          - unresolved
          - probable
        summary:
          type: string
          minLength: 1
          maxLength: 8192
        evidenceRefs:
          type: array
          maxItems: 100
          items:
            type: string
            format: uuid
        candidateUpdates:
          type: array
          maxItems: 50
          items:
            type: object
            additionalProperties: false
            properties:
              candidateKey:
                type: string
                pattern: ^sha256:[0-9a-f]{64}$
              reason:
                type: string
                minLength: 1
                maxLength: 4096
            required:
            - candidateKey
            - reason
        actionPlans:
          type: array
          maxItems: 10
          items:
            $ref: '#/components/schemas/ActionPlanV2'
        partial:
          type: boolean
        degradedSources:
          type: array
          maxItems: 100
          items:
            type: string
            maxLength: 128
      required:
      - schemaVersion
      - status
      - summary
      - evidenceRefs
      - candidateUpdates
      - actionPlans
      - partial
      - degradedSources
    InvestigationSuccessV2:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/InvestigationJobV2'
        requestId:
          type: string
      required:
      - data
      - requestId
      additionalProperties: false
    InvestigationStepsSuccessV2:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/InvestigationStepV2'
          maxItems: 128
        requestId:
          type: string
      required:
      - data
      - requestId
      additionalProperties: false
...
    CreateInvestigationV2:
      type: object
      properties:
        triggerRevision:
          type: integer
          minimum: 1
        triggerKind:
          enum:
          - manual
      required: []
      additionalProperties: false
    ActionPlanV2:
      type: object
      properties:
        schemaVersion:
          enum:
          - action-plan/v2
        actionPlanId:
          $ref: '#/components/schemas/SP06CommonIdentifier'
        tenantId:
          $ref: '#/components/schemas/SP06CommonTenantId'
        incidentId:
          $ref: '#/components/schemas/SP06CommonIdentifier'
        rcaRevision:
          type: integer
          minimum: 1
          maximum: 2147483647
        iteration:
          type: integer
          minimum: 1
          maximum: 2147483647
        targetCanonicalId:
          $ref: '#/components/schemas/SP06CommonCanonicalId'
        suggestedCommand:
          $ref: '#/components/schemas/SP06CommonCommandText'
        shell:
          enum:
          - bash
        recommendedExecutionProfileId:
          $ref: '#/components/schemas/SP06CommonIdentifier'
        reason:
          $ref: '#/components/schemas/SP06CommonText'
        expectedImpact:
          $ref: '#/components/schemas/SP06CommonText'
        preconditions:
          type: array
          items:
            $ref: '#/components/schemas/SP06CommonText'
          maxItems: 100
        riskSummary:
          $ref: '#/components/schemas/SP06CommonText'
        operationRisk:
          $ref: '#/components/schemas/SP06CommonOperationRisk'
        dataRisk:
          $ref: '#/components/schemas/SP06CommonDataClass'
        postCheckInstructions:
          type: array
          items:
            $ref: '#/components/schemas/SP06CommonText'
          maxItems: 100
        rollbackSuggestion:
          anyOf:
          - $ref: '#/components/schemas/SP06CommonText'
          - type: 'null'
        generatedBy:
          type: object
          properties:
            type:
              $ref: '#/components/schemas/SP06CommonShortText'
            name:
              $ref: '#/components/schemas/SP06CommonIdentifier'
            extensions:
              $ref: '#/components/schemas/SP06CommonExtensions'
          required:
          - type
          - name
          additionalProperties: false
        state:
          enum:
          - suggested
          - accepted
          - dismissed
          - superseded
        createdAt:
          $ref: '#/components/schemas/SP06CommonTimestamp'
        extensions:
          $ref: '#/components/schemas/SP06CommonExtensions'
      required:
      - schemaVersion
      - actionPlanId
      - tenantId
      - incidentId
      - rcaRevision
      - iteration
      - targetCanonicalId
      - suggestedCommand
      - shell
      - recommendedExecutionProfileId
      - reason
      - expectedImpact
      - preconditions
      - riskSummary
      - operationRisk
      - dataRisk
      - postCheckInstructions
      - rollbackSuggestion
      - generatedBy
      - state
      - createdAt
      additionalProperties: false
...
    SP06CommonTenantId:
      type: string
      minLength: 1
      maxLength: 128
      pattern: ^[A-Za-z0-9][A-Za-z0-9._-]*$
    SP06CommonIdentifier:
      type: string
      minLength: 1
      maxLength: 128
      pattern: ^[A-Za-z0-9][A-Za-z0-9._:-]*$
...
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  title: Smart Operations Platform API
  version: 1.3.0
...
openapi: 3.1.0
paths:
...
  /api/v1/incidents/{incidentId}/investigations:
    post:
      operationId: createInvestigation
      parameters:
      - in: path
        name: incidentId
        required: true
        schema:
          format: identifier
          type: string
      - description: Stable key for replay-safe write handling.
        in: header
        name: Idempotency-Key
        required: true
        schema:
          minLength: 1
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvestigationV2'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvestigationSuccessV2'
          description: Successful response.
        default:
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/ErrorEnvelope'
                - $ref: '#/components/schemas/SP04ErrorEnvelopeV2'
          description: Unified API error.
      security:
      - BearerAuth: []
      summary: createInvestigation
      tags:
      - Investigations
      x-idempotency: required
...
  /api/v1/investigations/{jobId}:
    get:
      operationId: getInvestigation
      parameters:
      - in: path
        name: jobId
        required: true
        schema:
          format: identifier
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvestigationSuccessV2'
          description: Successful response.
        default:
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/ErrorEnvelope'
                - $ref: '#/components/schemas/SP04ErrorEnvelopeV2'
          description: Unified API error.
      security:
      - BearerAuth: []
      summary: getInvestigation
      tags:
      - Investigations
...
  /api/v1/investigations/{jobId}/steps:
    get:
      operationId: getInvestigationSteps
      parameters:
      - in: path
        name: jobId
        required: true
        schema:
          format: identifier
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvestigationStepsSuccessV2'
          description: Successful response.
        default:
          content:
            application/json:
              schema:
                oneOf:
                - $ref: '#/components/schemas/ErrorEnvelope'
                - $ref: '#/components/schemas/SP04ErrorEnvelopeV2'
          description: Unified API error.
      security:
      - BearerAuth: []
      summary: getInvestigationSteps
      tags:
      - Investigations
...
security:
- BearerAuth: []
servers:
- url: /